second meandata refactoring, refs #12, #146
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@7995 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIEdge.cpp
@@ -252,7 +252,7 @@ GUIEdge::drawGL(const GUIVisualizationSettings &s) const throw() {
                 const SUMOReal length = segment->getLength();
                 if (idx < numCars.size()) {
                     const SUMOReal avgCarSize = segment->getOccupancy() / segment->getCarNumber();
-                    for (int i = 0; i < numCars[idx]; i++) {
+                    for (size_t i = 0; i < numCars[idx]; i++) {
                         SUMOReal vehiclePosition = position + length - i * avgCarSize;
                         SUMOReal xOff = 0.f;
                         while (vehiclePosition < position) {
